Added Controllers, Repositories, DBServices

+++ src/main/java/com/jobapplication/restgeoreceiver/domain/dto/GeolocationDto.java
@@ -11,8 +11,8 @@ import lombok.NoArgsConstructor;
 public class GeolocationDto {
 
     private Long id;
-    private Long latitude;
-    private Long longitude;
+    private double latitude;
+    private double longitude;
     private Device device;
 
 }
